// PIA Projekat/frontend/src/app/models/korisnik.ts
export class Korisnik{
    ime: string;
    prezime: string;
    korisnicko_ime: string;
    tip: string;
    lozinka: string;
    telefon: number;
    mejl: string;
    naziv_organizacije: string;
    drzava: string;
    grad: string;
    postanski_broj: number;
    ulica: string;
    broj: number;
    maticni_broj: number;
    status: string;
    privremena_lozinka: string;
    trenutak_zahteva: Date;
    slika: string;

    otvorenaNit: boolean;
}